pre_complete(del_api) 이메일 오류 발생, 이상한 코드

+++ lib/testcheck.dart
@@ -170,7 +170,7 @@ class _TestCheckScreenState extends State<TestCheckScreen> {
   }
 
   void navigateToEmailScreen(BuildContext context, List<ProblemData> dataList) {
-    final testPaperId = dataList[0].testPaperId as int;
+    final testPaperId = dataList[0].testPaperId;
 
     if (testPaperId != null) {
       Navigator.push(
